refactored "module" to "core"
