+++ src/main/java/dev/nero/aimassistance/core/TargetType.java
@@ -1,4 +1,4 @@
-package dev.nero.aimassistance.module;
+package dev.nero.aimassistance.core;
 
 public enum TargetType {
     ENTITY, BLOCK, NONE
